[Scene]: 'static-gradients' scene is moved to React 'StaticGradientScene' component.
Changelog:
- old features is restored in new 'scene' component;
- old 'scene.js' file is deleted now.
